+++ src/components/layouts/AvaliacaoBox.jsx
@@ -13,6 +13,7 @@ import { FiEye } from 'react-icons/fi';
 import { coresTags } from '../utils/coresTags';
 import ModalAvaliacao from '../ModalAvaliacao';
 import fotoDePerfilPadrao from '../../assets/fotoDePerfilPadrao.webp';
+import ModalAvaliacaoDetalhada from '../ModalAvaliacaoDetalhada';
 
 //Box de Avaliação, recebe uma avaliação e a apresenta em um perfil ou no Feed
 //Consegue verificar se o usuário logado é o autor da avaliação para exibir botões de edição e exclusão
@@ -23,6 +24,7 @@ export default function AvalicaoBox({ avaliacao, onChange }) {
   const { user } = useAuth();
   const op = useRef(null);
   const [modalVisivel, setModalVisivel] = useState(false);
+  const [modalAvaliacaoDetalhada, showModalDetalhar] = useState(false);
 
   const ehAutor = user.id === avaliacao.usuario.id;
 
@@ -83,7 +85,13 @@ export default function AvalicaoBox({ avaliacao, onChange }) {
         </button>
 
         <OverlayPanel className="border-[#1E1E1E] border bg-[#f7d799] text-[#1E1E1E]" ref={op}>
-          <button className="w-full flex items-center gap-2 px-6 py-2 hover:bg-[#e0b874] transition-colors duration-200">
+          <button
+            className="w-full flex items-center gap-2 px-6 py-2 hover:bg-[#e0b874] transition-colors duration-200"
+            onClick={e => {
+              showModalDetalhar(true);
+              op.current.hide(e);
+            }}
+          >
             <FiEye className="w-[32px] h-[32px]" />
             <p>Detalhar</p>
           </button>
@@ -173,6 +181,9 @@ export default function AvalicaoBox({ avaliacao, onChange }) {
           avaliacaoId={avaliacao.id}
           onEditSuccess={() => onChange && onChange()}
         />
+      )}
+      {modalAvaliacaoDetalhada && (
+        <ModalAvaliacaoDetalhada idAvaliacao={avaliacao.id} onClose={() => showModalDetalhar(false)} />
       )}
     </div>
   );